Fix exportAsHtml after bad merge into redux
